Ticket #2869 Refactor experiment page Experiment chart data moved out from the ExperimentResultAdapter - UpDownExpression class moved to the atlas-model
